refactor(manufacture): extract ConfirmProductCompletionWorkflow
Decompose the 102-line ConfirmProductCompletionAsync into explicit
steps (UpdateProducts, CalculateResidue, ThresholdGate, SubmitToErp,
UpdateBoM, TransitionToCompleted).

Behavior change: BoM update failures are now collected and appended to
the order note with ManualActionRequired=true, instead of silently
disappearing into warning logs.
